[Add] dynamic plot

+++ src/main/java/org/veritas/assessment/biz/dto/PlotFetchDto.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.NotNull;
 @Data
 public class PlotFetchDto {
     @NotNull
-    private Long modelArtifactVid;
+    private Integer modelArtifactVid;
 
     /**
      * <img id="" class="" src="">
